pyproject.toml
[tool.poetry]
name = "experiment"
version = "0.1.0"
description = ""
authors = ["RichardMathewsII <richardmathews.ai@gmail.com>"]
readme = "README.md"
packages = [{include = "experiment"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
mlflow = "^2.13.2"
dagster = "^1.7.9"
transformers = "^4.41.2"
langchain = "^0.2.3"
dagster-webserver = "^1.7.9"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
boto3 = "^1.34.122"
dagster-mlflow = "^0.23.9"
ipython = "^8.25.0"
ipykernel = "^6.29.4"
pymupdf = "^1.24.7"
tiktoken = "^0.7.0"
google-generativeai = "^0.7.1"
langchain_experimental = "^0.0.62"
langchain-openai = "^0.1.14"
psycopg2-binary = "^2.9.9"
streamlit = "^1.36.0"
tokencost = "^0.1.11"
nltk = "^3.8.1"

[tool.dagster]
module_name = "experiment.pipeline"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true
path = "."